Move platform-agnostic extensions to Sextant, introduce Sextant.Avalonia
