+++ src/Sextant.Avalonia/Sextant.Avalonia.csproj
@@ -1,0 +1,18 @@
+﻿<Project Sdk="MSBuild.Sdk.Extras">
+  <PropertyGroup>
+    <TargetFrameworks>netstandard2.0;net5.0</TargetFrameworks>
+    <AssemblyName>Sextant.Avalonia</AssemblyName>
+    <RootNamespace>Sextant.Avalonia</RootNamespace>
+    <PackageId>Sextant.Avalonia</PackageId>
+    <LangVersion>latest</LangVersion>
+  </PropertyGroup>
+  <ItemGroup>
+    <PackageReference Include="ReactiveUI" Version="13.*" />
+    <PackageReference Include="Avalonia" Version="0.10.0-rc1" />
+    <PackageReference Include="Avalonia.Desktop" Version="0.10.0-rc1" />
+    <PackageReference Include="Avalonia.ReactiveUI" Version="0.10.0-rc1" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\Sextant\Sextant.csproj" />
+  </ItemGroup>
+</Project>
